CTS test VCN Network Policies with Test Networks.
This CL adds CTS testing to make sure that Test Networks are
marked as VCN-managed when they have a subId that belongs to a
subscription group of a configured VCN.

This CL also defines a TestNetworkAgent for testing VCN APIs.
Specifically, this NetworkAgent is used to create a TUN Network
associated with a TestNetworkInterface. TestNetworkAgents will also
listen for VCN Network policies.

Bug: 182291467
Bug: 189125789
Test: atest FrameworksVcnTests CtsVcnTestCases
Change-Id: Ia7a256be4be949f1cd90ea98183990591a760f86