Enable secondary_user_on_secondary_display for CtsVcnTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 354264548

Test: atest --user-type secondary_user_on_secondary_display \
          CtsVcnTestCases
Test: atest CtsVcnTestCases
Change-Id: I5bee41272175af5e440889af4061b2c6cfbff919