bpf: remove unused tethering xdp programs
We can always trivially add it back if we ever manage
to make XDP work, but that will likely be on at least 6.1+
as that's the minimum kernel version on Pixel 6+ now.

In the mean time, we can just delete it, and very
slightly speed up boot time (less things to verify).

Note: code search in android for
  xdp_tether -file:kernel_logs/
  xdp/tether -file:kernel_logs/
doesn't find any hits outside this one file.

Test: TreeHugger
Flag: EXEMPT dead code removal
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I39c040d79b4f175cafc141e8f0c9e631e372ebd2