Wrap IEthernetServiceListener
This will be useful, so all callback related logic (such as filtering
based on USE_RESTRICTED_NETWORKS permission etc.) can be moved to
EthernetListener. In addition, it removes the try / catch blocks and
makes the code easier to read.

EthernetListener implements IInterface so it can continue to be stored
in the RemoveCallbackList inside EthernetTracker.

Test: TH
Flag: EXEMPT mainline
Change-Id: I86657d1681658d84fc2d9879a3dc5f47d8b705e4