Move restricted networks permission inside EthernetListener
Rather than using the cookie object. This means ListenerInfo can now be
removed.

Eventually, the ethernet code should not have to deal with this at all
internally. Instead, the InterfaceState object should be passed to
EthernetListener (instead of just the ifname) which will then figure out
by itself whether to send the callback. However, as long as the
tethering interface exists, this cannot be done.

The permission is also populated in removeListener(), though it is
entirely unused in that code path. Creating an EthernetListener object
in removeListener() is technically unnecessary anyway.  Unfortunately,
RemoteCallbackList#unregister() does not just take an IInterface object.
There are a few ways solutions I can think of:
a) Declare the RemoteCallbackList as
RemoteCallbackList<IEthernetServiceListener> and cast the listener
object to EthernetListener.
b) Create a wrapper class of RemoteCallbackList that does this under the
hood.
c) Store EthernetListener as the cookie, rather than the broadcast object.

Test: TH
Flag: EXEMPT mainline
Change-Id: I78fc68228f127ddee7d28e56499fc14a3aac1f73