Add previously released API to service-mdns-droidstubs
This is needed because Metalava is going to be strict about requiring a
previously released API to ensure that flagged APIs work correctly.
Module `service-mdns-droidstubs` does not use flagged APIs or provide
compatibility guarantees on its API but it is also using `droidstubs`
in a very different way to every other use in the build and it is not
worth special casing it.

Flag: EXEMPT build change to improve flag support
Bug: 425309457
Test: TH
Change-Id: I98797295693ff3a554db013e9c142f1a62dd6860