Add Support for TEMP_NOT_METERED for Carrier WiFi
Carrier WiFi will now support carrier networks going
from METERED->UNMETERED when they are on VCN. Because
the network is merged, we do not want to set the network
to be *actually* unmetered (it may become metered again),
which could cause undesired data usage; however, for
apps that opt in to support this additional info, and
for the system itself (including data accounting), we
want to enhance the user experience.

As part of this change, make sure that a couple other bits
are managed properly as well.

-Add support for TEMP_NOT_METERED to be passed through.
-Add support for NOT_BANDWIDTH_CONSTRAINED which appears to
 have been added without updating the VCN. This is a bug.
-Allow NOT_CONGESTED to be passed through rather than forced.
 This would allow the carrier to report the underlying network
 as congested to reduce unncessary traffic.

Bug: 436351169
Test: FrameworksVcnTests CtsVcnTestCases
Flag: EXEMPT BUGFIX
Change-Id: Ic7cc89c15b2e40cd31f27df40e4b8cac47b840b9