Add InCallService for OTT call detection
This change introduces the `ConnectivityCallListenerService`, an
InCallService that runs within the system_server process. Its purpose is
to detect the start and end of Over-the-Top (OTT) voice and video calls
made through the Telecom framework.

By binding to this service, the system can identify the UID of the
calling application and forward it to ConnectivityService. This enables
the platform to dynamically apply a high-priority network request for a
premium slice, enhancing the Quality of Service (QoS) for the call
without requiring app modifications.

This service is added as a MODULE_LIBRARIES component, which is the
optimal solution for performance and build integration as it runs within
the system_server process.

Flag: com.android.tethering.flags.enable_incall_service_api
Bug: 430509489
Bug: 445104509
Test: atest

Low-Coverage-Reason: ConnectivityCallListenerService is a skeleton implementation.

Change-Id: Idd4e0e1acfe3cd1e936ffbc21c4630e17d787a1c