Add onNoSocketCreated and onNetworkWithNoSocketDestroyed method in Socket Callbacks
onNoSocketCreated method will be invoked when socket creation did not succeed because
the network interface did not support mDNS. This will be used as a
signal to trigger offloaded scan to all the Offloaded Engines registered
on this particular network interface.

onNetworkWithNoSocketDestroyed will be invoked when a network for which
previously onNoSocketCreated was invoked is destroyed

Test: atest
Bug: 395515556
Flag: NONE This CL introduces new interface without any caller. The
relevant flag will be added in ag/35412901

Change-Id: I247f1fbccc99f1ef354fabf42c5b020a75db2a04