Invoke onNoSocketCreated and onNetworkWithNoSocketDestroyed if mDNS is not supported
Invoke onNoSocketCreated callback when socket creation request did not
succeed to create a socket. This is the trigger for creating receive
only mdns service type client.
Also ensure that onNetworkWithNoSocketDestroyed is invoked correctly
which will ensure proper cleanup.

Bug: 395515556
Test: atest ConnectivityCoverageTests:android.net.connectivity.com.android.server.connectivity.mdns.MdnsSocketProviderTest
Flag: NONE The code in this CL is behind the local compile time flag set
to false. This flag will be replaced with the Aflag in ag/35412901

Change-Id: I66e6d4eab27eecafd126c465ba76fd66698f5b3d