BpfUtils: Adjust API level reporting for 26Q2/trunk_staging.
Remap the calculated API level `3613` (which corresponds to `36.1.1 !REL` on `26Q2/trunk_staging`)
to `3700` to align with the expected future API level of this branch now that the `25Q4` branch point has passed.

Do it here, because this also becomes the NetBpfLoad version number,
and it's easier to do it here than to adjust the bpf_map_def.h header file.

Test: TreeHugger
Flag: EXEMPT UNSUPPORTED_LANGUAGE
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Iabe64f2c266a2a866c09d9bac0cf7cda2de4f07f