Add BPF map for uid/host/IPaddr access allowlist
Add a BPF map that allows allowlisting specific
uid/host/IP address combinations for local network access.

This is actually a trie, as maps always preallocate all the keys, which
would be wasteful for this use-case. The key holds most of the data
(uid, interface, IP address), and values are just booleans.

Entries can be added in particular when the user chose to
allow access to a particular device for an app (planned to
be implemented via NsdManager). They can be cleaned up
at least when the Network disconnects and its interface
is not tracked anymore.

Test: atest bpf_existence_test
Flag: NONE dead code
Change-Id: Iac80ba341d6762d3b4759bea42d92a54fb7dfd92