Revert "Disable address registration beta flag in Dhcp6Pd tests."
This reverts commit 309c3ebdd504591af81685891d3fac0c94eb0f53.

Reason for revert: This is not needed anymore, because address
registration is now only enabled if the RA includes M or O flags. This
also ended up making some of the tests really flaky, because the test is
configured in the constructor and cleaned up in tearDown. This means
that when the test is skipped tearDown is prevented from running.

Bug: 446103961
Change-Id: Iab95977615d1f15b805a71dee1cec48435f1585a