Do not call destroy() in finalizer
Calling destroy() in the finalizer is not a good idea because it calls
setIncludeTestInterfaces(false) which may break (or worse, flake)
subsequent tests.

Furthermore, when used in conjuction with AutoCloseTestInterfaceRule, the interface
itself will already have been cleaned up by the time this runs.

A better solution would be to make AutoCloseTestInterfaceRule a more
generic AutoCloseTestResourceRule, create a TestResource interface
(which is in essence very similar and might even inherit from
AutoCloseable) and register test resources with this rule. That way, the
AutoClose rule can clean up the resources directly.

Test: TH
Flag: NONE yes
Bug: 447704460
Change-Id: I2f5420c856e915f92c78278dc1e7aacf553663a1