Redesign AutoCloseTestInterfaceRule
AutoCloseTestInterfaceRule used to be a bit clumsy to use in combination
with EthernetTestInterface. This change makes AutoCloseTestInterfaceRule
much more generic, so that it can manage any kind of test resource that
implements AutoCloseable.

In the same change, a wrapper for TestNetworkInterface that implements
AutoCloseable is introduced, and EthernetTestInterface now also
implements AutoCloseable.

With this change, it should be harder to use EthernetTestInterface
incorrectly. Previously, it was still possible to do so, particularly
if the interface was allocated in the constructor and tests were skipped
(so deallocation in teardown did not run). With the new model, there are
no more reasons for explicit deallocation in tearDown.

Test: TH
Flag: NONE yes
Change-Id: I296d3220452621a5ff7e7c1685e80e5574834745