Move HandlerThread inside EthernetTestInterface
Moving the HandlerThread inside EthernetTestInterface is done in
preparation for adding a rule that cleans up EthernetTestInterfaces.
Since this rule would run after tearDown, it EthernetTestInterface
cannot use a handlerThread that is stopped during tearDown.

On a side note: as a next step, consider removing the handler thread
entirely. It is not entirely clear whether it is even needed, or whether
the callbacks could just run on the binder thread directly.

Test: TH
Flag: NONE yes
Change-Id: I6a0b9e66d6e500311db1f398fd18bea53adf7023