Remove StateMachine from NsdService
NsdService was a 1-state StateMachine, which is not useful and adds
indentation to all the code.

Replace it with a simple Handler.

This removes the message handling logging in the dumpsys logs, but these
were generally not useful and often duplicate with the other shared
logs.

This change is best reviewed with diff ignoring whitespace changes.

Test: atest
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie5ce80f440a605e74705e1c0520e57e9609ccf08