Add FLAG_NO_PICKER to DiscoveryRequest
The flag will be used to prevent usage of the picker and force an error
if the caller does not have permissions to discover all services.

In the future it could also be useful in combination with a
FLAG_INCLUDE_USER_APPROVED, to allow discovering only
previously-approved services.

Test: atest
Bug: 454248446
Flag: NONE dead code, code using it will be flagged
Change-Id: Ic4e0771a30a5305cc3a007c4a62a1b3cadd92705