Combine httpengine_impl and httpengine_api into a single target
There's no reason for those targets to be separate, as that introduces
alot of trouble when trying to depend on package-private classes from
the API. We usually never ran into problems in the past because our
API structure was wrong by defining everything as abstract and letting
the impl deal with it. However, the abstract classes should have never
been abstract in the first place which is what will be fixed in
a follow-up CL, but for the time being. this will mitigate the problem
of having to define abstract classes.

Moved the `java_defaults` target to `external/cronet` as it's closely
related to how we setup `httpengine` target.

Flag: NONE there's no code change, this is a mechanical refactor.
Bug: 448870925
Test: TH + m
Change-Id: I49f00e6ebaab7c420231cc240d3bffe729794b13