Adjust the Selective mDNS Response Flag Coverage
All code related to the selective mDNS response feature is
expected to be covered by a flag. The offload info collection and
update in notifyOffloadStart and updateOffloadInfo should only be
executed when the flag is enabled. Therefore, a flag check is
added to the methods so they are covered by a flag.

Tests are updated to enable the selective offload flag to ensure
the correct logic is exercised.

Bug: 446024482
Test: atest ConnectivityCoverageTests
Flag: com.android.tethering.mainline.beta.nsd_selective_mdns_response_offload
Change-Id: I4ef6c33804f156d44b9ffb69f44cdcdbfd36dc0b