Allow unregistering never-registered callbacks
NsdManager#stopServiceDiscovery, stopServiceResolution,
unregisterServiceInfoCallback throw if the listener is not registered.

This is not very convenient, and it also includes the case where
discovery failed to start (onStartDiscoveryFailed for example), where
the failure callback is posted to the app executor after removing
the service from the map.

This makes it very hard to use the stop method without a try/catch,
since discovery may have failed to start, but the app hasn't received
onStartDiscoveryFailed yet, so it doesn't know whether it's legal to
call stopServiceDiscovery at that point.

Also this behavior is problematic for the NsdManager picker, as
ideally the listener would be automatically unregistered once the user
picks a service, but then stopServiceDiscovery would crash if called
afterwards.

Just let the stop* methods fail silently if the callback isn't
registered. This is guarded with a compat flag to ensure there is no app
compat issues.

Bug: 450699352
Test: atest
Flag: EXEMPT BUGFIX
Change-Id: Id5f54786e126d9c35d5e17eb2502f343d8782cde