Handle call removal after app uninstall
Scenario handled:
Uninstall Scenario: This happens when an app is uninstalled mid-call. To solve this, we must avoid looking up the UID when the
call is removed, as the package no longer exists.

Fix: The solution is to cache the UID when the call is added and retrieve it from the cache when the call is removed.

Flag: EXEMPT BUGFIX
Bug: 448546376
Bug: 430509489
Test: atest

Change-Id: I9d5b91e7ebb68bf9f59c10c65660f6cf7381a4eb