Add app name to NSD picker intent
To let the user know which app is requesting the discovery, include the
app name in the picker intent, so it can be shown by the picker dialog
UI.

The logic is the same as what WifiNetworkFactory uses for its Wi-Fi
picker.

UI changes will be done separately.

Bug: 454248446
Test: atest
Flag: com.android.tethering.flags.nsd_service_picker
Change-Id: I47b1548f1e4ed4f9d4ab26786f41e2c1e72c0836