Add retry logic for Wi-Fi connection
This change introduces a retry mechanism for connecting to a
Wi-Fi hotspot in hotspot-relevant tests. The connectToWifiConfig
method will now attempt to connect up to 3 times if the initial
connection fails.This addresses transient failures that occur
when the SoftAP is not yet ready or when Wi-Fi scanning fails to
detect the hotspot in time.

Additionally, the expectSoftApCompleted() call has been removed
from the startWifiTethering() function. This removal is justified
because some legacy devices do not notify this callback, and the
new retry mechanism already covers the necessary SoftAP
completion time.

Bug: 448345079
Bug: 447394281
Test: atest CtsConnectivityMultiDevicesTestCases
Flag: EXEMPT TEST_ONLY
Change-Id: I391deb2d37a7f3c8db1d11dc3a290e74aab8af8d