Revert^2 "Ensure Thread network is never mDNS capable"
Reason for revert: Thread interfaces can't actually handle mDNS messages yet, and packets should not be sent as-is to the Thread network, which is what NsdService expects. NsdService expectations for Thread networks need to be designed/finalized.

Also there are checks for NET_CAPABILITY_LOCAL_NETWORK in MdnsSocketProvider which cause Thread networks to not be considered anyway.

Bug: 444384717
Change-Id: I4b36748b274bc105a5971b1da56654c51c71c3aa