Remove androidx dependency from nearby
The module builds, and tests pass fine without it.

This dependency makes it harder to integrate some tests as it causes
jarjar rules for androidx to be generated. If Nearby needs it in the
future, it could be added back disabling the jarjar rules for example,
but right now it looks like the dependency is not necessary.

Test: atest NearbyUnitTests NearbyIntegrationPrivilegedTests \
          NearbyIntegrationUntrustedTests CtsNearbyFastPairTestCases
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie5f2d2f5570f7ca5d74bc6935276c30d10134233