Add CTS tests for carrier-aware avoid bad Wi-Fi setting
This commit introduces CTS tests for the new carrier-aware "avoid bad
Wi-Fi" setting APIs in `ConnectivitySettingsManager`.

The `testSetAvoidUnvalidated` in `ConnectivityManagerTest` is updated to
conditionally use the new APIs (`setNetworkAvoidBadWifi(Context, subId, int)`)
if it's supported.
This ensures that the test coverage extends to both the existing global
setting and the new subscription-specific setting.

This also fixes the issue because running old test on B+ devices fails.
Bug: 434842865
Flag: EXEMPT TEST_ONLY
Test: atest CtsNetTestCases
Change-Id: I53efbfc59b5d623efc769dd617972c60c96184d1