Add NsdPicker dialog
Add a dialog that can be shown to present NsdServiceInfo services for
the user to choose from.

This will be used by NsdService to prompt the user to choose a service
for discovery.

Most of the logic in the dialog is currently empty, it will be added in
a followup change.

Test: flagged off code; atest with tests in next change
Bug: 454248446
Flag: com.android.tethering.flags.nsd_service_picker
Change-Id: I14dbb61c48cdf1c37a4ebe90c1df1edd4e71054c