Split UnitTestsLib, FrameworksNetTests manifests
Use separate AndroidManifest files for ConnectivityUnitTestsLib (which
is included in ConnectivityCoverageTests) and FrameworksNetTests.

The two targets run different tests which require different permissions
and activity declarations, so it does not make sense for them to use the
same manifests. In fact, FrameworksNetTests requires almost no
permissions.

Test: atest ConnectivityCoverageTests FrameworksNetTests
Flag: EXEMPT TEST_ONLY
Change-Id: Ib5ed303424f6e237747f7914c8a5b8479e2b1026