Propagate cached capabilities bits to MdnsServiceInfo
This is a no-op refactoring to prepare for subsequent changes.

The cached capabilities bits in SocketKey are a snapshot of the
NetworkCapabilities at the time the socket is created. This information
is useful for clients of mDNS to know about the network they are
communicating on without having to query ConnectivityManager again.

This change ensures that the cached capabilities bits are passed from
the SocketKey, where they are originally stored, to the
MdnsServiceInfo when a service is discovered. The propagated
information will be used in subsequent changes.

Flag: com.android.tethering.flags.nsd_use_network_callback_for_local_networks
Test: atest ConnectivityCoverageTests --test-filter='.*Mdns.*' --update-device
Test: atest ConnectivityCoverageTests --test-filter='.*Nsd.*' --update-device
Test: atest CtsNetTestCases --test-filter='.*Nsd.*' --update-device
Test: m service-connectivity-mdns-standalone-build-test
Bug: 452297472
Change-Id: Ia060742acfe4b2785867369173ebcdeeca5c03d1