Move reusable test constants and helper methods into a new DnsSvcbTestUtils class
The eventual DnsHttpsPacketTest will want to reuse some of these constants and helper methods, so this change moves some of them into a new Kotlin companion object.

A follow-up CL may be done to move the TestDnsRecordByteArrayBuilder as this is a bit more complex (and might require Kotlin-izing DnsSvcbPacketTest so a data class can be used).

Bug: 454544870
Test: atest NetworkStaticLibTests
Flag: EXEMPT TEST_ONLY
Change-Id: I7df2abe9364d2c728a2caed12b43de6e422e0147