Refactor MdnsFeatureFlags Builder for Robustness
To prevent newly added feature flags from being overlooked in
NsdService, this change introduces a build-time check in the
MdnsFeatureFlags.Builder. The builder now ensures that all
feature flags are explicitly set before the build() method is
called. If any flag is not set, an IllegalStateException is
thrown. To accommodate flags that are used in legacy code but not
in NsdService, a default exemption list has been added.

Additionally, a setAllFlagsForTesting() method has been added to
the builder to simplify test setup.

Fix: 452299126
Test: atest ConnectivityCoverageTests --test-filter Nsd*
Test: atest ConnectivityCoverageTests --test-filter mdns*
Test: atest CtsNetTestCases --test-filter Nsd*
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ic8c18c492a095dd0d4faecfc36494d7837ae3479