Consolidate L2capNetwork callbacks into one
This change does not change behavior.

Both onError and onNetworkUnwanted ended up with the same
implementation, so there is no reason to distinguish the two. One thing
to note is the BluetoothSocket#close() triggers an exception on the Read
/ Write threads which is also propagated back via this callback, so it
is possible that the callback is called more than once. This shouldn't
be a problem because the server network checks whether the network still
exists in mL2capNetworks and similarly,
declareAllNetworkRequestsUnfulfillable(), which is called from the
client network callback, is also safe to call multiple times.

Test: TH
Flag: NONE yes
Change-Id: Ic4d89c451bf6cf751b662a6c4de9e0ceefc83fc6