Add JSR305 dependency to service-connectivity-pre-jarjar
Apparently, this dependency was being pulled from HttpEngine's through
static libs, when we started jarjaring HttpEngine at a different
depth, the compilation started failing as the dependency was hidden
from the rest of the modules which is working as intended.

This dependency should be added, regardless of whether HttpEngine
shadows JSR305 or not.

Bug: 458670023
Flag: EXEMPT PURE_REFACTOR
Change-Id: I5e880ba6d1263db506187af9cf6a958d640466b4