Change ETH_REQUEST to filter out physical interfaces
On automotive devices, physical interfaces that were satisfying
ETH_REQUEST were being interpreted in the tests in listeners and
callbacks, leading to many test failures.

This change alters ETH_REQUEST to be more selective towards the test
interfaces created. Transports currently match as an OR, as opposed to
how capabilities are treated as an AND. Therefore, the previous request
could match either Ethernet or Test transports, or networks that had
both Ethernet or Test transports. This is undesirable, and now, only
test transports will be matched. Additionally, NOT_METERED is one of the
few capabilities contained in test interfaces but not physical.

Test: atest EthernetManagerTest with eth0 configured
Bug: 448846480
Flag: EXEMPT bugfix
Change-Id: Idfef6bf0ce5841ca4dc7c4d8e956dce846afab3d