Use package name if app name is empty
Instead of showing an empty string to the user if the package is not
found or the app name empty, replace it with the package name. This is
not very user-readable but it is at least better than an empty string,
and can give the user an idea of what app it might be.

In practice this code path should not be used unless there is a race
where the app gets uninstalled during a request (in which case the
request will get cancelled due to the app kill), or the app actually has
an empty name (distribution channels will often not allow it).

Bug: 454248446
Test: atest
Flag: com.android.tethering.flags.nsd_service_picker
Change-Id: I03435de82d7edb781d53c55f58072bbb155d0428