Support concurrent intents in picker activity
Add a queue of intents so that if an intent is received while the picker
is already shown, it is queued to be used next.

This should allow supporting cases where apps file multiple discovery
requests, or multiple apps file a discovery request.

As part of this change onDismiss code in NsdPickerFragment is moved to
onDetach, so that the host activity can be notified when the picker is
fully done (the fragment is detached) and is able to show the next
dialog (attach the next fragment). Note onDismiss runs before onDetach
so the fragment would still be attached at that point.

Test: atest NsdPickerActivityTest
Bug: 454248446
Flag: com.android.tethering.flags.nsd_service_picker
Change-Id: Iac34d5d5f9b306336ec552eb6ac08096a09d54c9