bpf/netd: (5.15+ T+) allow netd (daemon) to bind to port 53
Now that DnsResolver has merged '25Q2+: stop using socketbroker' the
functionality of this change can be tested by reverting sepolicy change:
  netd.te: allow netd to bind to ports <1024 (including dns: 53 & 853)
ie. by changing //system/sepolicy/private/netd.te:
-allow netd self:global_capability_class_set { net_admin net_raw net_bind_service kill };
+allow netd self:global_capability_class_set { net_admin net_raw kill };

Which would normally cause the non-socketbroker code path to fail,
but cuttlefish is on 5.15+ kernel and thus (with this patch) it still works.

This can be tested via:
  atest 'ResolverMultinetworkTest#TetheringDnsForwardingTest'

In a follow up I'll also add a test for this to BpfHandler startup code.

Test: TreeHugger
Flag: EXEMPT UNSUPPORTED_LANGUAGE
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Icb0b0b81680f2d807f18d1fac93524a8ccf4f111