Prevent addPerAppDefaultNetworkRequests from updating default requests.
The addPerAppDefaultNetworkRequests() method is responsible for updating app-filed TRACK_DEFAULT requests when a per-app default network changes. It calls getPerAppCallbackRequestsToUpdate() to find
which requests need to be updated.

Before this change, getPerAppCallbackRequestsToUpdate() was
incorrectly including in its results requests that were present in
mDefaultNetworkRequests (i.e.,mDefaultRequest, and other per-app
default requests). This is incorrect because these requests are not
filed by apps and should not be modified. This caused addPerAppDefaultNetworkRequests() modify the default network requests
themselves.

This bug only affects satellite and only affects UID 1000 because:
- The code only considers requests that have a TRACK_DEFAULT as
  the first layer and currently only the satellite fallback request is
  structured in this way.
- The code only modifies requests filed by UIDs that are subject to
  a per-app default. Currently, multilayer requests are always filed
  by UID 1000. So the incorrect behaviour only happens if UID 1000 is
  subject to a per-app default.

This change fixes the issue skipping default network requests in
getPerAppCallbackRequestsToUpdate(). This ensures the method only
returns app-filed requests for updating, as intended.

Test: TH
Fix: 464433511
Flag: EXEMPT BUGFIX
Change-Id: Ied6eb80208871f0f8a21a4a60a21c8a5dd062583