Only override relevant VCN carrier config keys in VCN CTS
This is to fix the test failures on user builds when overriding
satellite carrier config is no longer allowed.

Previously, the test will cache the entire carrier config bundle and
reset the carrier config in the end by passing this entire
bundle to the override method. This starts to break when overriding
satellite keys are no long allowed. This is also unnecessary and
expensive because it causes the system to go through all the keys
that are irrelevant.

Bug: 465899199
Flag: EXEMPT TEST_ONLY
Test: atest CtsVcnTestCases:VcnManagerTest
Change-Id: I9285a0d635ea297b227131b52c22756313a51f43