Ensure tasks are removed from Handler in RealtimeScheduler
When removing a delayed runnable or message in RealtimeScheduler,
currently only the internal task queue is checked. However, if
the task has already been posted to the Handler (e.g. because the
delay expired or was 0), it would still execute.

This CL updates removeDelayedRunnable and removeDelayedMessage to
also explicitly remove the callbacks/messages from the underlying
Handler.

Bug: 457801989
Test: atest NetworkStaticLibTests:com.android.net.moduletests\
      .util.RealtimeSchedulerTest
Flag: EXEMPT BUGFIX
Change-Id: I510b1c46a97f0f7416c40b9607a99d6a39bc9ef1