Wait for messages posted after waitForIdle calls
waitForIdle posts a message and waits for it to be processed, but if
other messages are posted after that (for example by the code running on
the handler) the handler will still not be idle after that message has
been processed.

Actually wait for the handler to be idle using addIdleHandler on the
handler queue.

This is necessary for some upcoming tests, and may fix some flakes.

Test: Tests using this class in presubmit
Flag: EXEMPT TEST_ONLY
Change-Id: Id1a6a3461205cba9bdc7b183cccd829d9e586b53