Support caching for loopback events
Because there may be many loopback access events (especially in cases
where access is allowed), we want to rate limit the events that we put
in the event ring buffer. Per {src_uid, dst_uid, result}, only store one
event per minute. We'll keep track of the last event time by storing the
boot time as the value in the cache map.

Note this function is not used anywhere yet, but its use will be gating on the value in the `loopback_access_metrics_enabled_map`, which will not be populated until ag/37172580 is submitted.

Test: Treehugger
Bug: 431786207
Flag: com.android.tethering.flags.loopback_access_metrics
Change-Id: I81346089c3b7787b1a0ab4e7d258a025c3e623f2