UID comparison for loopback packets
On egress, when IP traffic is detected on the loopback interface,
perform a socket lookup on the destination 5-tuple in order to compare
the source and destination socket UIDs.

For TCP, we only do the lookup if we receive a SYN packet. If we're
receiving non-SYN TCP packets, we know the connection has already been
established and can avoid doing checks. For UDP, we have to do a lookup for
every packet. We'll consider adding caching to improve the UDP
performance as a follow up.

Test: Treehugger
Bug: 431786207
Flag: com.android.tethering.flags.loopback_access_metrics
Change-Id: I0b363d32b69bd87bea082fe33cb2cb348315047f